+++ Library/OcConsoleLib/UgaPassThrough.c
@@ -18,6 +18,8 @@
 #include <Library/MemoryAllocationLib.h>
 #include <Library/UefiBootServicesTableLib.h>
 
+#include <Library/OcMiscLib.h>
+
 STATIC
 EFI_STATUS
 EFIAPI
@@ -170,20 +172,9 @@ OcProvideUgaPassThrough (
   // and 1 UGA protocol:
   // - for unknown handle
   //
-  Status = gBS->LocateHandleBuffer (
-    ByProtocol,
-    &gEfiUgaDrawProtocolGuid,
-    NULL,
-    &HandleCount,
-    &HandleBuffer
-    );
-
-  if (!EFI_ERROR (Status)) {
-    DEBUG ((DEBUG_INFO, "OCC: Found %u handles with UGA draw\n", (UINT32) HandleCount));
-    FreePool (HandleBuffer);
-  } else {
-    DEBUG ((DEBUG_INFO, "OCC: Found NO handles with UGA draw\n"));
-  }
+  DEBUG_CODE_BEGIN ();
+  DEBUG ((DEBUG_INFO, "OCC: Found %u handles with UGA draw\n", (UINT32) OcCountProtocolInstances (&gEfiUgaDrawProtocolGuid)));
+  DEBUG_CODE_END ();
 
   Status = gBS->LocateHandleBuffer (
     ByProtocol,
@@ -192,62 +183,68 @@ OcProvideUgaPassThrough (
     &HandleCount,
     &HandleBuffer
     );
+  if (EFI_ERROR (Status)) {
+    DEBUG ((DEBUG_INFO, "OCC: Failed to find handles with GOP - %r\n", Status));
+    FreePool (HandleBuffer);
 
-  if (!EFI_ERROR (Status)) {
-    DEBUG ((DEBUG_INFO, "OCC: Found %u handles with GOP for UGA check\n", (UINT32) HandleCount));
+    return Status;
+  }
 
-    for (Index = 0; Index < HandleCount; ++Index) {
-      DEBUG ((DEBUG_INFO, "OCC: Trying handle %u - %p\n", (UINT32) Index, HandleBuffer[Index]));
+  DEBUG ((DEBUG_INFO, "OCC: Found %u handles with GOP for UGA check\n", (UINT32) HandleCount));
+  for (Index = 0; Index < HandleCount; ++Index) {
+    DEBUG ((DEBUG_INFO, "OCC: Trying handle %u - %p\n", (UINT32) Index, HandleBuffer[Index]));
 
-      Status = gBS->HandleProtocol (
-        HandleBuffer[Index],
-        &gEfiGraphicsOutputProtocolGuid,
-        (VOID **) &GraphicsOutput
-        );
-
-      if (EFI_ERROR (Status)) {
-        DEBUG ((DEBUG_INFO, "OCC: No GOP protocol - %r\n", Status));
-        continue;
-      }
-
-      Status = gBS->HandleProtocol (
-        HandleBuffer[Index],
-        &gEfiUgaDrawProtocolGuid,
-        (VOID **) &UgaDraw
-        );
-
-      if (EFI_ERROR (Status)) {
-        DEBUG ((DEBUG_INFO, "OCC: No UGA protocol - %r\n", Status));
-
-        OcUgaDraw = AllocateZeroPool (sizeof (*OcUgaDraw));
-        if (OcUgaDraw == NULL) {
-          DEBUG ((DEBUG_INFO, "OCC: Failed to allocate UGA protocol\n"));
-          continue;
-        }
-
-        OcUgaDraw->GraphicsOutput = GraphicsOutput;
-        OcUgaDraw->Uga.GetMode = OcUgaDrawGetMode;
-        OcUgaDraw->Uga.SetMode = OcUgaDrawSetMode;
-        OcUgaDraw->Uga.Blt = OcUgaDrawBlt;
-
-        Status = gBS->InstallMultipleProtocolInterfaces (
-          &HandleBuffer[Index],
-          &gEfiUgaDrawProtocolGuid,
-          &OcUgaDraw->Uga,
-          NULL
-          );
-
-        DEBUG ((DEBUG_INFO, "OCC: Installed UGA protocol - %r\n", Status));
-      } else {
-        DEBUG ((DEBUG_INFO, "OCC: Has UGA protocol, skip\n"));
-        continue;
-      }
+    Status = gBS->HandleProtocol (
+      HandleBuffer[Index],
+      &gEfiGraphicsOutputProtocolGuid,
+      (VOID **) &GraphicsOutput
+      );
+    if (EFI_ERROR (Status)) {
+      DEBUG ((DEBUG_INFO, "OCC: No GOP protocol - %r\n", Status));
+      continue;
     }
 
-    FreePool (HandleBuffer);
-  } else {
-    DEBUG ((DEBUG_INFO, "OCC: Failed to find handles with GOP\n"));
+    Status = gBS->HandleProtocol (
+      HandleBuffer[Index],
+      &gEfiUgaDrawProtocolGuid,
+      (VOID **) &UgaDraw
+      );
+    if (!EFI_ERROR (Status)) {
+      DEBUG ((DEBUG_INFO, "OCC: Skipping UGA proxying as it is already present on handle %u - %p\n", (UINT32) Index, HandleBuffer[Index]));
+      continue;
+    }
+
+    OcUgaDraw = AllocateZeroPool (sizeof (*OcUgaDraw));
+    if (OcUgaDraw == NULL) {
+      DEBUG ((DEBUG_INFO, "OCC: Failed to allocate UGA protocol\n"));
+      continue;
+    }
+
+    OcUgaDraw->GraphicsOutput = GraphicsOutput;
+    OcUgaDraw->Uga.GetMode = OcUgaDrawGetMode;
+    OcUgaDraw->Uga.SetMode = OcUgaDrawSetMode;
+    OcUgaDraw->Uga.Blt = OcUgaDrawBlt;
+
+    Status = gBS->InstallMultipleProtocolInterfaces (
+      &HandleBuffer[Index],
+      &gEfiUgaDrawProtocolGuid,
+      &OcUgaDraw->Uga,
+      NULL
+      );
+    if (EFI_ERROR (Status)) {
+      FreePool (OcUgaDraw);
+    }
+
+    DEBUG ((
+      DEBUG_INFO,
+      "OCC: Installed UGA protocol - %r (Handle %u - %p)\n",
+      Status,
+      (UINT32) Index,
+      HandleBuffer[Index]
+      ));
   }
+
+  FreePool (HandleBuffer);
 
   return Status;
 }