+++ Changelog.md
@@ -13,6 +13,7 @@ OpenCore Changelog
 - Fixed switching to graphics mode when entering OpenCanopy
 - Fixed installing Apple FB Info protocol when no GOP exists
 - Fixed abort timeout sound in OpenCanopy on key press
+- Added `GopPassThrough` option to support GOP protocol over UGA
 
 #### v0.6.6
 - Added keyboard and pointer entry scroll support in OpenCanopy